+++ main.lua
@@ -15,7 +15,7 @@ if love and (type(mod) ~= "table" or type(mod.read) ~= "function") then
   return require("viewer.App")
 end
 
-mod.exports.version = "1.1.1"
+mod.exports.version = "1.1.2"
 
 local namespace = { mod = mod, path = mod.path }
 local modules = {}